+++ shared/fs/common/index.js
@@ -1,0 +1,5 @@
+// @flow
+export {default as PathItemAction} from './path-item-action-container'
+export {default as OpenInSystemFileManager} from './open-in-system-file-manager-container'
+export {default as PathItemInfo} from './path-item-info'
+export {default as PathItemIcon} from './path-item-icon'
