+++ shared/fs/filepreview/default-view.js
@@ -4,8 +4,7 @@ import * as Types from '../../constants/types/fs'
 import * as Constants from '../../constants/fs'
 import {globalStyles, globalColors, globalMargins, platformStyles} from '../../styles'
 import {Box, Button, Text} from '../../common-adapters'
-import PathItemInfo from '../common/path-item-info'
-import PathItemIcon from '../common/path-item-icon'
+import {PathItemInfo, PathItemIcon} from '../common'
 import {memoize} from 'lodash-es'
 import {fileUIName, isMobile, isIOS} from '../../constants/platform'
