+++ shared/fs/row/editing.js
@@ -4,7 +4,7 @@ import * as Types from '../../constants/types/fs'
 import * as Styles from '../../styles'
 import {rowStyles} from './common'
 import {Icon, ClickableBox, Input, Box, Button, Text} from '../../common-adapters'
-import PathItemIcon from '../common/path-item-icon'
+import {PathItemIcon} from '../common'
 
 type EditingProps = {
   name: string,
